requirements.txt
streamlit 
python-dotenv
requests